moving retries around
